% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbv_get_parameters.R
\name{hbv_get_parameters}
\alias{hbv_get_parameters}
\title{Download Global HBV Parameter Maps}
\usage{
hbv_get_parameters(aoi, folds = 1:10, mean = FALSE, warp = TRUE)
}
\arguments{
\item{aoi}{SpatVector. A polygon layer with area of interest.}

\item{folds}{Numeric. A numeric vector from 1 to 10 indicating the
folds numbers to process.}

\item{mean}{Logical. If TRUE, return mean zonal statistics,
calculated using \code{global} method from \code{terra} package}

\item{warp}{Logical. If TRUE, reproject the HBV rasters to
\code{aoi} projection}
}
\value{
List
}
\description{
This function download Global HBV Parameter Maps (v0.8)
created by \emph{Beck et al.} (2020) and crop it to the shapefile
boundaries.
}
\examples{

# Load shapefile
f <- system.file("ex/lux.shp", package="terra")
v <- vect(f)

# Get zonal statisitcs
fold1_mean <- hbv_get_parameters(v, folds = 1, mean = TRUE, warp = FALSE)
fold1_mean

# Get rasters
fold1_raster <- hbv_get_parameters(v, folds = 1, mean = FALSE, warp = FALSE)
plot(fold1_raster[[1]])

}
\references{
Beck HE, Pan M, Lin P, Seibert J, van Dijk AIJM, Wood EF. 2020. Global Fully Distributed Parameter Regionalization Based on Observed Streamflow From 4,229 Headwater Catchments. Journal of Geophysical Research: Atmospheres 125 : e2019JD031485. DOI: 10.1029/2019JD031485
}
