% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbv_get_parameters.R
\name{hbv_get_parameters}
\alias{hbv_get_parameters}
\title{Download Global HBV Parameter Maps}
\usage{
hbv_get_parameters(
  aoi,
  folds = 1:10,
  mean = FALSE,
  warp = TRUE,
  version = "v0.9"
)
}
\arguments{
\item{aoi}{SpatVector. A polygon layer with area of interest.}

\item{folds}{Numeric. A numeric vector from 1 to 10 indicating the
folds numbers to process.}

\item{mean}{Logical. If TRUE, return mean zonal statistics,
calculated using \code{\link[terra]{global}} method from \pkg{terra} package}

\item{warp}{Logical. If TRUE, reproject the HBV rasters to
\code{aoi} projection}

\item{version}{Character. Can be used to determine the dataset version. Must be
one of "v0.8" (original release from February 5, 2020) or "v0.9" (revised version
from May 5, 2020). See details.}
}
\value{
List
}
\description{
This function download Global HBV Parameter Maps
created by \emph{Beck et al.} (\href{https://agupubs.onlinelibrary.wiley.com/doi/10.1029/2019JD031485}{2020})
and crop it to the shapefile boundaries. See
\href{http://www.gloh2o.org/hbv/}{GloH2O website} for details.
}
\details{
Version history (according to \href{http://www.gloh2o.org/hbv/}{GloH2O website}):

\describe{
  \item{V0.9 (May 5, 2022)}{To avoid local minima, the authors reduced the number
  of predictors to three and increased the number of generations to 2000. A selection
  was made based on which predictors individually provided the best training score:
  snowfall fraction of precipitation (\code{FSNOW}), mean topographic slope (\code{SLOPE}),
  and soil clay content (\code{CLAY}). Increasing the number of generations to 2000
  allowed the algorithm to find the true optimum, while lowering the spatial
  resolution from 0.05° to 0.1° reduced computational requirements.}
  \item{V0.8 (February 5, 2020)}{Original release corresponding to \emph{Beck et al.}
  (\href{https://agupubs.onlinelibrary.wiley.com/doi/10.1029/2019JD031485}{2020}).}
}
}
\examples{

# Load shapefile
f <- system.file("ex/lux.shp", package="terra")
v <- vect(f)

# Get zonal statisitcs
fold1_mean <- hbv_get_parameters(v, folds = 1, mean = TRUE, warp = FALSE)
fold1_mean

# Get rasters
fold1_raster <- hbv_get_parameters(v, folds = 1, mean = FALSE, warp = FALSE)
plot(fold1_raster[[1]])

}
\references{
Beck HE, Pan M, Lin P, Seibert J, van Dijk AIJM, Wood EF. 2020. Global Fully Distributed Parameter Regionalization Based on Observed Streamflow From 4,229 Headwater Catchments. Journal of Geophysical Research: Atmospheres 125 : e2019JD031485. DOI: 10.1029/2019JD031485
}
